Moved around memory map again for RGB Color support
