+++ lib/audio/audio.js
@@ -14,7 +14,7 @@ const DEFAULT_AUDIO_LATENCY_IN_MILLI = 100;
 const WASMBOY_MIN_TIME_REMAINING_IN_MILLI = 75;
 
 // https://docs.google.com/spreadsheets/d/17xrEzJk5-sCB9J2mMJcVnzhbE-XH_NvczVSQH9OHvRk/edit?usp=sharing
-const WASMBOY_MEMORY_SOUND_INDEX = 0x05B800;
+const WASMBOY_MEMORY_SOUND_INDEX = 0x0B2000;
 const WASMBOY_SAMPLE_RATE = 48000;
 
 // Some canstants that use the ones above that will allow for faster performance
