+++ lib/memory/memory.js
@@ -9,7 +9,7 @@ import { idbKeyval } from './idb';
 // Declare some memory constants
 const MEMORY_ADDRESSES = {
   CARTRIDGE_RAM: 0x010400,
-  CARTRIDGE_ROM: 0x07B800,
+  CARTRIDGE_ROM: 0x0D2000,
   WASMBOY_INTERNAL_STATE: {
     LOCATION: 0x000000,
     SIZE: 0x000400
